Correctly manage files that should be flushed to disk and somehow refreshed before the model update.
Currently changing a ceylon config file in an Idea editor triggers a ful build while the Idea document is not yet saved to disk.
So the triggered model update takes in account the previous content of the config file found on disk.

For the fix of #390 (Changes in .ceylon/* are not used unless the project is reopened)